refactor: drain ShortLeaseDetails into useShortLeaseDetails composable
Move the ShortLeaseDetails.vue script logic (quote-derived computeds,
liquidation math, and the debounced Skip swap-fee fetch) into a
useShortLeaseDetails composable, leaving the SFC as a template over the
returned bindings. Behaviour and template are unchanged; the 600ms swap-fee
debounce and single onUnmounted timer teardown move verbatim.

Brings the SFC from 446 to 168 LOC (issue #185).
